[SQL] Add default SQL syntax selector command
This commit adds a naive command for users to choose default SQL dialect
to be assigned to `source.sql`.

The `sql_set_default_syntax` basically works for any syntax, but is prefixed
`sql_` as it is located in and dedicated to SQL syntax package.

Alternatively a general purpose `set_default_syntax` could be implemented
by ST core or Default package.

Ideally, dialect syntaxes could be identified by attributes in `Syntax()` class,
but until this becomes true, rely on a static list of files specified via
`dialect_file` argument to choose from.